+++ lib/core/bulk_downloads/src/pages/bulk_download_page.dart
@@ -6,8 +6,6 @@ import 'package:kurumi/material.dart';
 import 'package:material_symbols_icons/symbols.dart';
 
 // Project imports:
-import '../../../boorus/booru/types.dart';
-import '../../../configs/config/providers.dart';
 import '../../../widgets/widgets.dart';
 import '../providers/bulk_download_notifier.dart';
 import '../providers/saved_download_task_provider.dart';
@@ -20,21 +18,7 @@ class BulkDownloadPage extends ConsumerWidget {
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    final config = ref.watchConfigAuth;
-
-    return config.booruType == BooruType.zerochan
-        ? Scaffold(
-            appBar: AppBar(
-              title: Text(context.t.sideMenu.bulk_download),
-            ),
-            body: Center(
-              child: Text(
-                'Temporarily disabled due to an issue with getting the download link'
-                    .hc,
-              ),
-            ),
-          )
-        : const BulkDownloadPageInternal();
+    return const BulkDownloadPageInternal();
   }
 }