+++ packages/booru_clients/lib/src/zerochan/types/types.dart
@@ -1,4 +1,5 @@
 export 'autocomplete_dto.dart';
+export 'download_urls.dart';
 export 'enums.dart';
 export 'post_dto.dart';
 export 'tag_dto.dart';